eciespy_test/requirements.txt
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.3.0
bleach==3.1.5
Brlapi==0.6.6
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.1
clickclick==1.2.2
coincurve==13.0.0
command-not-found==0.3
connexion==2.6.0
cryptography==2.1.4
cupshelpers==1.0
cytoolz==0.10.1
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.16
eciespy==0.3.6
eth-hash==0.2.0
eth-keys==0.3.3
eth-typing==2.2.1
eth-utils==1.9.0
filelock==3.0.12
Flask==1.1.2
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.1
importlib-resources==1.4.0
inflection==0.4.0
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.2
jsonschema==3.2.0
keyring==21.2.1
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
MarkupSafe==1.1.1
meld==3.18.0
netifaces==0.10.4
nose==1.3.7
oauth==1.0.1
olefile==0.45.1
openapi-spec-validator==0.2.8
packaging==20.3
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.13.1
protobuf==3.0.0
py==1.8.1
pybitflyer==0.1.9
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
pycryptodomex==3.9.8
pycups==1.9.73
Pygments==2.6.1
pygobject==3.26.1
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.3
python-bitbankcc==0.0.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==5.3.1
readme-renderer==26.0
reportlab==3.4.0
requests==2.23.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
SecretStorage==3.1.2
simplejson==3.13.2
six==1.14.0
swagger-ui-bundle==0.0.6
system-service==0.3
systemd-python==234
toml==0.10.0
toolz==0.10.0
tox==3.14.6
tqdm==4.46.1
twine==3.1.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.16
virtualenv-clone==0.5.4
wadllib==1.3.2
webencodings==0.5.1
Werkzeug==1.0.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
